todo-app/requirements.txt
flask==2.3.2
pymongo==4.3.3
flask-cors==4.0.0
python-dotenv==1.0.0